requirements.txt
pandas>=2.0.0
streamlit>=1.28.0
plotly>=5.15.0
scikit-learn>=1.3.0
numpy>=1.24.0
pyjanitor>=0.25.0
featuretools>=1.27.0
xlsxwriter>=3.1.0
setuptools>=68.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
